+++ _example/request-binding/main.go
@@ -240,5 +240,5 @@ func main() {
 	log.Println("     -F 'file=@/path/to/file.pdf'")
 	log.Println()
 
-	log.Fatal(app.Graceful(":8080"))
+	log.Fatal(app.Start(":8080"))
 }